+++ addons/account_peppol_selfbilling/models/__init__.py
@@ -1,4 +1,3 @@
-from . import account_edi_proxy_user
 from . import account_edi_xml_ubl_bis3
 from . import account_journal
 from . import account_move_send
